Refactor: Move OptionsPopup to own package
